Create generic slide component
A generic slide component has been added which accepts a "text" property
and renders it as Markdown.

A reactive data store has also been configured along with an initially
simple interval for looping through slides. The mechanism for advancing
the slide handles slides being added or removed from the slides array,
as well as the order of slides changing. This includes the current slide
being removed or changing position.