Consume API in front-end app
The app entrypoint now calls the API to fetch the slides to display. An
API helper file has been added with functions to seed and generate the
API token, as well as to make the API call and return the data.

The type definitions have been updated to reflect the data that is
returned from the API.

There is currently no refresh mechinism so the slides are currently only
loaded once when the app is initialised.