Send slide updated events from server
When slides are added, removed, or modified, an event will be sent from
the server via Mercure. A Doctrine event subscriber listens for
postPersist, postRemove, and postUpdate lifecycle events and publishes
an event to the Mercure hub.

In the prod environment, the Mercure events are sent as private,
requiring authentication to listen to them. In dev and test
environments, the events are sent as public to allow unauthenticated
clients to listen them. This is done to circumvent cross-origin issues
during development and testing.

The Symfony web profiler has also been added which provides a way to
monitor Mercure events sent by the application.