Consume slide updated events in frontend app
When an event is received from the Mercure hub, the app will update the
slides collection in the data store. This covers slides being added,
removed, and updated.

The base Twig template has an aditional script tag above the main app
scripts to set the Mercure hub URL as a window variable. This is done so
that the Twig mercure function can be used to generate the URL and set
the authentication cookie. The authentication cookie is always set in
the frontend app but will only be used in the prod environment, it will
be ignored in dev and test environments.