Add Mercure environment variables to workflow
Three environment variables have been added to the deployment workflow
to configure the Mercure bundle for the slide updated events.

MERCURE_JWT_SECRET is the secret for signing the tokens used for
authentication, MERCURE_URL is the URL used by the server to publish
events to the Mercure hub, MERCURE_PUBLIC_URL is the URL used by the
frontend app to subscribe to the events.