Refresh page after Mercure disconnection
The JWT authentication token has an expiry time of one hour. After this,
the Mercure event source would disconnect due to a 401 HTTP response. To
overcome this, a whole page refresh is triggered in the event of a
connection error with the event source.

The page refresh will only be triggered following a previous successful
connection. This is to prevent a constant refresh if the Mercure hub is
not reachable on initial page load.